revamp working queryablestate
