+++ Features/queryablestate-carinsurance/clients/src/main/java/net/corda/samples/carinsurance/Client.java
@@ -1,0 +1,36 @@
+package net.corda.samples.carinsurance;
+
+import net.corda.client.rpc.CordaRPCClient;
+import net.corda.core.messaging.CordaRPCOps;
+import net.corda.core.node.NodeInfo;
+import net.corda.core.utilities.NetworkHostAndPort;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.List;
+
+import static net.corda.core.utilities.NetworkHostAndPort.parse;
+
+/**
+ * Connects to a Corda node via RPC and performs RPC operations on the node.
+ *
+ * The RPC connection is configured using command line arguments.
+ */
+public class Client {
+    private static final Logger logger = LoggerFactory.getLogger(Client.class);
+
+    public static void main(String[] args) {
+        // Create an RPC connection to the node.
+        if (args.length != 3) throw new IllegalArgumentException("Usage: Client <node address> <rpc username> <rpc password>");
+        final NetworkHostAndPort nodeAddress = parse(args[0]);
+        final String rpcUsername = args[1];
+        final String rpcPassword = args[2];
+        final CordaRPCClient client = new CordaRPCClient(nodeAddress);
+        final CordaRPCOps proxy = client.start(rpcUsername, rpcPassword).getProxy();
+
+        // Interact with the node.
+        // For example, here we print the nodes on the network.
+        final List<NodeInfo> nodes = proxy.networkMapSnapshot();
+        logger.info("{}", nodes);
+    }
+}
