+++ Features/queryablestate-carinsurance/contracts/build.gradle
@@ -5,7 +5,7 @@ cordapp {
     targetPlatformVersion corda_platform_version.toInteger()
     minimumPlatformVersion corda_platform_version.toInteger()
     contract {
-        name "carinsurance Contract"
+        name "Queryablestate Car Insurance"
         vendor "Corda Open Source"
         licence "Apache License, Version 2.0"
         versionId 1
@@ -46,5 +46,4 @@ dependencies {
     cordaRuntime "$corda_release_group:corda:$corda_release_version"
     compileOnly "$corda_release_group:corda-testserver-impl:$corda_release_version"
     testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-
 }
