--- Features/queryablestate-carinsurance/contracts/src/main/java/net/corda/examples/carinsurance/schema/InsuranceSchemaFamily.java
@@ -1,9 +1,0 @@
-package net.corda.examples.carinsurance.schema;
-
-/**
- * Schema Family for Insurance Mapped Schema
- *
- * MappedSchema should be associated with a schema family which is consistent across versions of the schema.
- * It allows the SchemaService to select the appropriate version of the schema if it has evolved over time.
- */
-public class InsuranceSchemaFamily { }
