+++ Features/queryablestate-carinsurance/gradle.properties
@@ -1,3 +1,3 @@
-name=Test
+name=Queryablestate Car Insurance
 group=com.carinsurance
 version=0.1
