+++ Features/queryablestate-carinsurance/workflows/build.gradle
@@ -6,7 +6,7 @@ cordapp {
     targetPlatformVersion corda_platform_version
     minimumPlatformVersion corda_platform_version
     workflow {
-        name "carinsurance Flows"
+        name "Queryablestate Car Insurance"
         vendor "Corda Open Source"
         licence "Apache License, Version 2.0"
         versionId 1
